Implemented exclusions by post/category for tag and series lists (fixed #37)
